add a surround with snippet command, https://github.com/microsoft/vscode/issues/86397
